# No defer in inc perf