After a bit profiling I see that incEmitted and incCollected is still hot, I saw that defer takes most of the time, after reading a bit it looks that defer has a bit of overhead and in small functions/hot code isn't recommended (should be checked in benchmark before checking, this is not a rule)

```
benchmark                              old ns/op        new ns/op        delta
Benchmark_T1000_P5000_Matches-4        10894368013      9009168906       -17.30%
Benchmark_T1000_P5000_NoMatches-4      11772957621      31820886         -99.73%
Benchmark_T100_P5000_Matches-4         1149160968       684346858        -40.45%
Benchmark_T100_P5000_NoMatches-4       1193448589       25386986         -97.87%
Benchmark_T10_P5000_Matches-4          189739122        127694632        -32.70%
Benchmark_T10_P5000_NoMatches-4        155396278        23437142         -84.92%
Benchmark_T10_P500_CountTask-4         20287106         18338299         -9.61%
Benchmark_T10_P50000_CountTask-4       1832485835       1721740236       -6.04%
Benchmark_T1000_P500-4                 2034353343       1993059158       -2.03%
Benchmark_T1000_P50000_CountTask-4     197003231081     202667675345     +2.88%

benchmark                              old allocs     new allocs     delta
Benchmark_T1000_P5000_Matches-4        69144          69126          -0.03%
Benchmark_T1000_P5000_NoMatches-4      69146          65069          -5.90%
Benchmark_T100_P5000_Matches-4         56523          56502          -0.04%
Benchmark_T100_P5000_NoMatches-4       56548          56075          -0.84%
Benchmark_T10_P5000_Matches-4          55226          55212          -0.03%
Benchmark_T10_P5000_NoMatches-4        55219          55165          -0.10%
Benchmark_T10_P500_CountTask-4         10940          10936          -0.04%
Benchmark_T10_P50000_CountTask-4       1154692        1154676        -0.00%
Benchmark_T1000_P500-4                 542557         543978         +0.26%
Benchmark_T1000_P50000_CountTask-4     61019034       61027137       +0.01%

benchmark                              old bytes      new bytes      delta
Benchmark_T1000_P5000_Matches-4        4477632        4474896        -0.06%
Benchmark_T1000_P5000_NoMatches-4      4476112        3959169        -11.55%
Benchmark_T100_P5000_Matches-4         3668616        3667368        -0.03%
Benchmark_T100_P5000_NoMatches-4       3671944        3614036        -1.58%
Benchmark_T10_P5000_Matches-4          3585190        3584450        -0.02%
Benchmark_T10_P5000_NoMatches-4        3584604        3578712        -0.16%
Benchmark_T10_P500_CountTask-4         1624756        1624493        -0.02%
Benchmark_T10_P50000_CountTask-4       114281120      114280304      -0.00%
Benchmark_T1000_P500-4                 128125336      128215768      +0.07%
Benchmark_T1000_P50000_CountTask-4     7625275048     7625790528     +0.01%
```
_Diff between master and this branch, defaultEdgeBufferSize = 0_